FAC-965:impl:Move CreateSubjectInsightsHelper to insight.subject package
