+++ flickit-assessment-core/src/main/java/org/flickit/assessment/core/application/service/insight/subject/InitSubjectInsightService.java
@@ -7,10 +7,10 @@ import org.flickit.assessment.common.exception.AccessDeniedException;
 import org.flickit.assessment.common.exception.ResourceNotFoundException;
 import org.flickit.assessment.core.application.port.in.insight.subject.InitSubjectInsightUseCase;
 import org.flickit.assessment.core.application.port.out.assessmentresult.LoadAssessmentResultPort;
-import org.flickit.assessment.core.application.port.out.subject.LoadSubjectPort;
 import org.flickit.assessment.core.application.port.out.insight.subject.CreateSubjectInsightPort;
 import org.flickit.assessment.core.application.port.out.insight.subject.LoadSubjectInsightPort;
 import org.flickit.assessment.core.application.port.out.insight.subject.UpdateSubjectInsightPort;
+import org.flickit.assessment.core.application.port.out.subject.LoadSubjectPort;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
